avanzo test, falta arreglar el tiempo de espera

+++ src/main/java/com/modelos_y_simulacion_2024/dominio/Behavior.java
@@ -2,5 +2,5 @@ package com.modelos_y_simulacion_2024.dominio;
 
 @FunctionalInterface
 public interface Behavior {
-  int nextTime();
+  float nextTime();
 }
