+++ src/main/java/com/modelos_y_simulacion_2024/dominio/Event.java
@@ -1,19 +1,19 @@
 package com.modelos_y_simulacion_2024.dominio;
 
 public final class Event {
-  private final int clock;
+  private final float clock;
   private final Behavior behavior;
   private final Entidad enditad;
   private final Planificator planificator;
 
-  public Event(int clock ,Behavior behavior, Entidad entidad, Planificator planificator ) {
+  public Event(float clock ,Behavior behavior, Entidad entidad, Planificator planificator ) {
     this.clock = clock;
     this.behavior = behavior;
     this.enditad = entidad;
     this.planificator = planificator;
   }
 
-  public int getClock() {
+  public float getClock() {
     return clock;
   }
