+++ src/main/java/com/modelos_y_simulacion_2024/escenario/Behaviors/TestBehavior.java
@@ -4,14 +4,14 @@ import com.modelos_y_simulacion_2024.dominio.Behavior;
 
 public class TestBehavior implements Behavior{
 
-  private final int fixedClock;
+  private final float fixedClock;
 
-  public TestBehavior(int fixedClock) {
+  public TestBehavior(float fixedClock) {
     this.fixedClock = fixedClock;
   }
 
   @Override
-  public int nextTime() {
+  public float nextTime() {
     return this.fixedClock;
   }
